Reorganize vcx agency wallet plugin loading code
