+++ vcx/dummy-cloud-agent/src/indy/mod.rs
@@ -3,7 +3,6 @@ pub mod did;
 pub mod logger;
 pub mod pairwise;
 pub mod wallet;
+pub mod wallet_plugin;
 pub use indyrs::{ErrorCode, IndyError};
-
-
